# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simanja (Sistem Management Jamaah)",
        "description": "This is developped using Swagger UI.  You can find\r\nout more about Swagger at\r\n[http://swagger.io](http://swagger.io) or on\r\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "iqbalhanif313@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/setup/kategori-jamaah": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Jenis Kegiatan information",
                "operationId": "setup/kategori-jamaah",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "setup/kategori-jamaah"
                ],
                "summary": "Store Setup KategoriJamaah",
                "operationId": "setup/kategori-jamaah/store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Store Setup KategoriJamaah",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Store Setup KategoriJamaah",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kategori-jamaah/{id}": {
            "get": {
                "tags": [
                    "setup/kategori-jamaah"
                ],
                "summary": "Show Detail Setup KategoriJamaah",
                "operationId": "setup/kategori-jamaah/{id}",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "setup/kategori-jamaah"
                ],
                "summary": "Update Setup KategoriJamaah",
                "operationId": "setup/kategori-jamaah/update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Update Setup KategoriJamaah",
                                        "type": "string",
                                        "example": "DRH"
                                    },
                                    "nama": {
                                        "description": "Update Setup KategoriJamaah",
                                        "type": "string",
                                        "example": "Daerah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "setup/kategori-jamaah"
                ],
                "summary": "Delete Setup KategoriJamaah",
                "operationId": "setup/kategori-jamaah/{id}/delete",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kecamatan": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kecamatan information",
                "operationId": "setup/kecamatan",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kecamatan/{st_kab_id}": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kecamatan information",
                "operationId": "setup/kecamatan/id",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kelurahan": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kelurahan information",
                "operationId": "setup/kelurahan",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kelurahan/{st_kab_id}": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kelurahan information",
                "operationId": "setup/kelurahan/id",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}